Move class GameState() to game_state.py
